requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-CORS==4.0.0
openai==1.3.0
dashscope==1.14.0
requests==2.31.0
python-dotenv==1.0.0
supabase==2.0.0
websocket-client==1.6.4
